templates: don't make missing template a fatal error
Notably, a packet could contain data, template, data, template and when
this is the case, the template will never be parsed.

There is also a slight API change: we return data and error only for
non-fatal errors (missing template) and not for other errors (where we
return nil and error). This still leaves a chance for the user to report
the missing templates and this is less confusing as when we can get both
data and error and when not.

sflow decoder is also modified with the same logic (return nil and error
instead of data and error since all errors should be considered fatal).

See https://github.com/akvorado/akvorado/issues/697